% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALLI.v0.2.R
\docType{class}
\name{LargeDataObject-class}
\alias{LargeDataObject-class}
\title{Class LargeDataObject
Class \code{LargeDataObject} holds large data such as technical variance and results from BALLI fit}
\description{
Class LargeDataObject
Class \code{LargeDataObject} holds large data such as technical variance and results from BALLI fit
}
