% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALLI.v0.2.R
\docType{class}
\name{TecVarList-class}
\alias{TecVarList-class}
\title{Class TecVarList
Class \code{TecVarList} holds technical variance}
\description{
Class TecVarList
Class \code{TecVarList} holds technical variance
}
